[Refactor] 라벨 로딩을 싱글톤 LabelLoader로 분리 - 기존 ImageClassifier에서 라벨 로딩 기능을 분리하여 LabelLoader (싱글톤)으로 이동 - LabelLoader가 앱 실행 중 한 번만 JSON을 로드하고 캐싱하도록 변경 - ImageClassifier는 이미지 분류 기능만 담당하도록 역할을 명확히 분리 - 성능 최적화: 불필요한 JSON 로드 방지 및 메모리 사용 개선 - 유지보수성 향상: 라벨 로딩을 별도 모듈로 분리하여 확장 가능성 증가 (다른 AI 분류 모델로 확장 시 같은 라벨 활용 가능)
